+++ TODO.txt
@@ -5,7 +5,7 @@
 #5. В классе Preparator сделать один публичный и остальные приватные
 #6. тоже самое в остальных классах
 7. сделать отдельно файл генераторы
-8. check_word - ожидается булево, а он закрывает, делим на чек и терминейт
+#8. check_word - ожидается булево, а он закрывает, делим на чек и терминейт
 9. функциональность препейр и клиер вывести в декораторы
 10. отделить бизнес-логику от тулов
 11. сделать отдельный класс под юзкейс
